Enemy HP bar revision
changed from slider to fill bar

+++ PP2GameDevelopment/Assets/Scripts/Enemy/EnemyAI.cs
@@ -134,7 +134,7 @@ public class EnemyAI : MonoBehaviour, IDamage
         hp -= amount;
         agent.SetDestination(gameManager.instance.player.transform.position);
         StartCoroutine(FlashDamage());
-        healthBar.UpdateHealthBar(hp, hpOrig);
+        healthBar.UpdateHealthBar((float)hp, hpOrig);
 
         if (hp <= 0)
         {
